+++ app/.gitignore
@@ -37,3 +37,6 @@ next-env.d.ts
 
 # .env
 .env
+
+# node_modules
+node_modules
