+++ app/package.json
@@ -15,7 +15,8 @@
     "next": "14.2.3",
     "openai": "^4.47.3",
     "react": "^18",
-    "react-dom": "^18"
+    "react-dom": "^18",
+    "voicevox-api-client": "^1.1.1"
   },
   "devDependencies": {
     "@types/node": "^20",
