+++ app/src/types/voicevox-api-client.d.ts
@@ -1,0 +1,1 @@
+declare module "voicevox-api-client";
