+++ docs/voice-vox/index.md
@@ -1,0 +1,8 @@
+## VOICE VOX Test
+
+```bash
+curl -X 'POST' \
+  'http://localhost:50021/audio_query?text=pyupyumaru&speaker=1' \
+  -H 'accept: application/json' \
+  -d ''
+```
