+++ voice-vox/README.md
@@ -1,0 +1,11 @@
+# VOICE VOX 環境
+
+1. VOICE VOX を使って、音声合成、音声出力を実施する
+
+## 参考・引用
+
+1. [VOICEVOX GitHub](https://github.com/VOICEVOX)
+
+2. [VoiceVox で高クオリティな音声合成をかんたんに行う](https://zenn.dev/ddpn08/articles/fb390caa781e54)
+
+3. [VOICEVOX(音声合成)を REST-API で利用する](https://qiita.com/A_T_B/items/1531d78944d8b796b9fa)
