+++ voice-vox/docker-compose.yml
@@ -1,0 +1,19 @@
+version: "2"
+services:
+  voicevox_engine:
+    # CPUの場合は voicevox/voicevox_engine:cpu-ubuntu20.04-latest を使用
+    image: voicevox/voicevox_engine:cpu-ubuntu20.04-latest
+    # image: voicevox/voicevox_engine:nvidia-ubuntu20.04-latest
+    ports:
+      - "50021:50021"
+    tty: true
+    #
+    # 以下の項目はCPUの場合はなくても大丈夫
+    #
+    # deploy:
+    #   resources:
+    #     reservations:
+    #       devices:
+    #         - driver: nvidia
+    #           count: 1
+    #           capabilities: [gpu]
