+++ voice-vox/package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "voicevox-api-client": "^1.1.1"
+  }
+}
